+++ scripts/vars.sh
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2020-2023 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2020-2024 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -24,10 +24,11 @@
 
 export PRODUCT_CSM="csm"
 export PRODUCT_COS="cos"
-export VERSION="csm-1.5"
+CSM_MAJ_MIN="1.6"
+export VERSION="csm-${CSM_MAJ_MIN}"
 
 export SLES_VERSION="15"
-export SLES_SPNUM=5
+export SLES_SPNUM="6"
 export SLES_SP="SP${SLES_SPNUM}"
 export SLES_SP_LOWER="sp${SLES_SPNUM}"
 export SLES_ARCH="x86_64"
@@ -45,7 +46,7 @@ then
     export CSM_RELEASE_VERSION="master"
     export COS_RELEASE_VERSION="master"
 else
-    export CSM_RELEASE_VERSION="1.5"
+    export CSM_RELEASE_VERSION="${CSM_MAJ_MIN}"
     export COS_RELEASE_VERSION="2.2"
 fi