+++ CHANGELOG.md
@@ -6,10 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- CASMINST-7001: Update CSM release to 1.6 and SLES SP to 6
 
 ## [2.5.2] - 2024-07-26
 ### Added
-- CASMCMS-9066: Add `procps` to `bootstrap` package list in `kiwi-ng/cray-sles15sp5-barebones/config-template.xml.j2`,
+- CASMCMS-9066: Add `procps` to `bootstrap` package list in `kiwi-ng/cray-sles15sp6-barebones/config-template.xml.j2`,
   to fix build failure.
 
 ## [2.5.1] - 2024-04-24